requirements_versions.txt
torch
accelerate>=0.15.0
albumentations>=1.3.0
basicsr>=1.4.2
clean_fid
einops>=0.6.0
facexlib
fonts>=0.0.3
font-roboto
imageio
jsonmerge>=1.9.0
kornia>=0.6.8
lark>=1.1.5
matplotlib>=3.6.2
numexpr
omegaconf>=2.3.0
opencv_python_headless
packaging
pandas>=1.5.2
piexif>=1.1.3
pip
psutil>=5.9.4
pudb
Pygments
pytorch_lightning==1.4.2
torchmetrics==0.6.0
pytz>=2022.7
regex
resize_right
scikit_image
setuptools>=65.5.0
timm>=0.6.12
tqdm>=4.64.1
wandb>=0.13.7
pyside6>=6.4.1
numpy>=1.23.1
pillow>=9.2.0
pyyaml>=6.0
opencv-python>=4.6.0.66
clip>=1.0
scipy>=1.9.3
ftfy
future
gfpgan>=1.3.5
realesrgan>=0.3.0
scikit-image>=0.19.3
requests>=2.28.1
gitdb
typing_extensions
cffi
attrs
wheel
cryptography
rsa
pyasn1
wcwidth
six
atomicwrites
pyinstaller
lmdb>=1.4.0
urwid
colorama
yapf
idna
multidict
filelock
smmap
lpips>=0.1.4
numba
llvmlite
scikit-learn>=1.2.0
fsspec
tensorboard
urllib3>=1.26.13
click
shortuuid
boto3
pathtools
gevent
promise
joblib
addict
networkx
aiohttp
yarl
cachetools
certifi
imgaug
markupsafe
python-dateutil
pyarrow
aiosignal
frozenlist
chardet
tifffile
shiboken6
filterpy
markdown
oauthlib
win_inet_pton
werkzeug
contourpy
fonttools
jsonschema
pyparsing
pyrsistent
cycler
kiwisolver
botocore
bottle
tokenizers
pydantic>=1.10.2
fairscale>=0.4.13
k-samplers
open_clip_torch
torchsde>=0.2.5
fastapi>=0.88.0
uvicorn>=0.20.0
pydantic
transformers>=4.25.1
diffusers>=0.10.2
safetensors>=0.2.6
k-diffusion
xformers
torchvision>=0.13.1
gdown>=4.6.0
torchdiffeq>=0.2.3
fire>=0.5.0
test-tube
torchaudio==0.12.1
pedalboard
pydub
